test: isolate model router process fixture branches
